+++ Modules/Intent.Modules.HotChocolate.GraphQL/Templates/MutationType/MutationTypeTemplateRegistration.cs
@@ -7,7 +7,7 @@ using Intent.Modelers.Services.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.Registrations;
 using Intent.Modules.HotChocolate.GraphQL.Models;
-using Intent.Modules.Modelers.Services.GraphQL.Api;
+using Intent.Modelers.Services.GraphQL.Api;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 using GraphQLMutationTypeModel = Intent.Modules.HotChocolate.GraphQL.Models.GraphQLMutationTypeModel;