+++ Modules/Intent.Modules.HotChocolate.GraphQL/Templates/QueryType/QueryTypeTemplatePartial.cs
@@ -11,7 +11,7 @@ using Intent.Modules.Common.CSharp.TypeResolvers;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Constants;
 using Intent.Modules.HotChocolate.GraphQL.Models;
-using Intent.Modules.Modelers.Services.GraphQL.Api;
+using Intent.Modelers.Services.GraphQL.Api;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 using Intent.Utils;
@@ -29,6 +29,7 @@ namespace Intent.Modules.HotChocolate.GraphQL.Templates.QueryType
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public QueryTypeTemplate(IOutputTarget outputTarget, IGraphQLQueryTypeModel model) : base(TemplateId, outputTarget, model)
         {
+            AddNugetDependency(NuGetPackages.HotChocolate);
             SetDefaultCollectionFormatter(CSharpCollectionFormatter.Create("System.Collections.Generic.IReadOnlyList<{0}>"));
             AddTypeSource(TemplateFulfillingRoles.Application.Contracts.Dto);
             AddTypeSource(TemplateFulfillingRoles.Domain.Entity.Primary);