+++ Modules/Intent.Modules.HotChocolate.GraphQL/Templates/SubscriptionType/SubscriptionTypeTemplatePartial.cs
@@ -6,7 +6,7 @@ using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Constants;
-using Intent.Modules.Modelers.Services.GraphQL.Api;
+using Intent.Modelers.Services.GraphQL.Api;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
@@ -23,6 +23,7 @@ namespace Intent.Modules.HotChocolate.GraphQL.Templates.SubscriptionType
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public SubscriptionTypeTemplate(IOutputTarget outputTarget, GraphQLSubscriptionTypeModel model) : base(TemplateId, outputTarget, model)
         {
+            AddNugetDependency(NuGetPackages.HotChocolate);
             AddTypeSource(TemplateFulfillingRoles.Application.Contracts.Dto);
             AddTypeSource(TemplateFulfillingRoles.Domain.Entity.Primary);
             AddTypeSource(TemplateFulfillingRoles.Domain.ValueObject);