+++ Tests/GraphQL.CQRS.TestApplication/GraphQL.CQRS.TestApplication.Application/Interfaces/IProductsService.cs
@@ -12,15 +12,10 @@ namespace GraphQL.CQRS.TestApplication.Application.Interfaces
 
     public interface IProductsService : IDisposable
     {
-
         Task<Guid> CreateProduct(ProductCreateDto dto);
-
         Task<ProductDto> FindProductById(Guid id);
-
         Task<List<ProductDto>> FindProducts();
-
         Task UpdateProduct(Guid id, ProductUpdateDto dto);
-
         Task DeleteProduct(Guid id);
 
     }